Move Object.create(provides) from provide to component creation
Due to the way Muban components are created and when their setup
function is called, re-creating an inherited version when provide
is called is not the write time, since all client provides are
already linked up. This does work in Vue because components
and their setup functions are called "per component" instead of
"per operation".

That's why we've moved the provides creation to the Component itself,
and pro-actively create a new "clone" of the provides object for
each component, so it can be added to at any later time.