Improve bindMap and other bindings after bindTemplate updates
Previously there were some memory leaks regarding bindMap when
the HTML was updated with bindTemplate. New bindings were created
to make everything work, but old bindings were not removed.

The fix here is done in two ways:

1) The `getRefs` for collections how returns an array or refs that
are managed by the collection itself, so when the collection changes,
these refs can be set to `undefined`, so any bindings to this ref
will be cleaned up.

2) Additionally, bindMap now cleans up its own bindings, and registers
it watcher to the Component instance, so both gets cleaned up.