Make the props object readonly
The props object passed to the setup function inside a component
contains information from the outside, either from the HTML (initially),
or from the parent component.

The direction of information is meant to be one-way, so writing back to props
should not be used. It already never communicated back to the HTML or parent
component, so it could only be used for "internal state".

While this internal state sounds useful, it can also be confusing. Making this
readonly makes it clear that props should only be used for one thing. If you
try to write to it, it will log a warning in the console.