Check current HTML value before setting innerHTML
Setting innerHTML (even if the value is the same) triggers
MutationObserver checks to see if either new components should be
initialized, or if existing components should be unmounted.

Since reading the current DOM is much cheaper compared to the writing +
side-effects, this should improve things.

The commented out code was a simple "batching" implementation that seems
to benefit slightly in synthetic benchmarks, but since its current
implementation is not robust enough, it's currently still disabled.