Proxy watch/watchEffect to allow auto-cleanup on component unmount
`watch` and `watchEffect` exist in runtime-core instead of reactivity
packages because they are tied to the Vue components, and their
lifecycles. Which has some advantages, that Muban didn't have.

With this update, we proxy those two functions to allow for auto-cleanup
when the Muban component unmounts. It will also ignore usages when
invoked in an unmounted component.

For library consumers this should work out of the box, as long as they
use the muban exports (and not the vue exports directly).