Correctly clean up old refs in collections
Previously when, the DOM updates, and a collection is "refreshed", it
would assume that all elements are new, and set all old refs to
`undefined`.

However, when appending in the DOM manually, elements (and their
component instances) can be "moved" from the old to the new collection.
By setting them to `undefined` in the old collection, they would also be
`undefined` in the new collection.