Make lazy type-safe and add runtime checking
The imported file from `getComponent` now must have an export name that
is either the `displayName` in `PascalCase`, or the `exportName`. This
is enforced in TS, and shows a mismatch error if that's not the case.

If you wanted to, you could still choose to ignore the types somehow,
resulting in a runtime error. That scenario is now properly checked, and
shows a clear error message.