Improve CSS binding story tests
Refactor story component and story for a cleaner code.
Move the html template to a function and use 'refCollection' instead of 'ref' to bind the checkboxes to reduce code.
Add a locked class to the box element for testing purposes.
Replace previous 'Story' with 'Story' in the CFSv3 (object) format.
Add interaction testing using the 'play' function for a more useful test.
Remove eslint disable rule comment.