Simplify lazy component usage
Instead of relying on the `lazy` export convention, determine the `export` name
based on the component `displayName` that is already passed as the first parameter.
On mismatch,  allow explicitly providing the `exportName`.

The imported file from `getComponent` now must have an export name that
is either the `displayName` in `PascalCase`, or the `exportName`. This
is enforced in TS, and shows a mismatch error if that's not the case.

If you wanted to, you could still choose to ignore the types somehow,
resulting in a runtime error. That scenario is now properly checked, and
shows a clear error message.

Add stories and update docs and old example code.