Also export declarations in `dist/esm` folder
Previously it was not possible to import anything from
`@muban/muban/dist/esm/...` directly, since it wouldn't match the path
for types; `@muban/muban/types/...`.

With this change, the declarations are now also exported in the
`dist/esm` folder, so that those files can be imported without any TS
error.

The existing `types` folder is still there for convenience in case
non-exported types should be imported.